Changed double to decimal to fix rounding error

+++ Program.cs
@@ -1,5 +1,5 @@
 ﻿
 Console.Write("Enter two numbers: ");
-double a = Convert.ToDouble(Console.ReadLine());
-double b = Convert.ToDouble(Console.ReadLine());
+decimal a = Convert.ToDecimal(Console.ReadLine());
+decimal b = Convert.ToDecimal(Console.ReadLine());
 Console.WriteLine($"The sum of {a} and {b} is {(a+b)}");
